Add changes from feedback by @opyh @mutaphysis
