+++ src/components/Link/Link.js
@@ -1,31 +1,8 @@
 // @flow
 import React, { createContext, Component } from 'react';
 
-import {
-  type RouterHistory,
-  type RouterHistoryMethod,
-  type RouteParams,
-} from '../lib/RouterHistory';
-
-type RouteContext = {
-  params: RouteParams,
-  history: RouterHistory,
-  name: string,
-};
-
-const { Provider: RouteProvider, Consumer: BaseRouteConsumer } = createContext<?RouteContext>();
-
-const RouteConsumer = ({ children }: { children: (context: RouteContext) => React$Node }) => (
-  <BaseRouteConsumer>
-    {(context: ?RouteContext) => {
-      if (context == null) {
-        throw new Error('Missing route context.');
-      }
-
-      return children(context);
-    }}
-  </BaseRouteConsumer>
-);
+import { type RouterHistory, type RouteParams } from '../../lib/RouterHistory';
+import { type RouteContext, RouteConsumer, RouteProvider } from './RouteContext';
 
 type Props = {
   routeName: string,
@@ -42,7 +19,9 @@ class Link extends Component<Props> {
   ) {
     const { routeName, replace, onClick } = this.props;
 
-    onClick && onClick(event);
+    if (onClick) {
+      onClick(event);
+    }
 
     if (event.isDefaultPrevented()) {
       return;
@@ -60,7 +39,7 @@ class Link extends Component<Props> {
   renderLink = (context: RouteContext) => {
     const { history } = context;
     const { routeName, params, ...props } = this.props;
-    const href = history.generate(routeName, params);
+    const href = history.generatePath(routeName, params);
 
     delete props.replace;
